add test create_new_house_CC
