+++ src/test/java/javaTests/pageObjects/SpinnerPage.java
@@ -18,9 +18,9 @@ public class SpinnerPage extends DriverWait {
     public MobileElement currentHouseName;
     @AndroidFindBy(xpath = "//android.widget.TextView[@text='Войти по приглашению']")
     public MobileElement enterByInviteBtn;
-    @AndroidFindBy(xpath = "//android.widget.TextView[@text='CC Test_Name']")
+    @AndroidFindBy(xpath = "//android.widget.TextView[@text='CC test']")
     public MobileElement houseNameCC;
-    @AndroidFindBy(xpath = "//android.widget.TextView[@text='Wi-fi Test_Name']")
+    @AndroidFindBy(xpath = "//android.widget.TextView[@text='Wi-fi test']")
     public MobileElement houseNameWIFI;
 
 }
