+++ src/test/java/javaTests/tests/CreateNewHouseWifiTest.java
@@ -1,21 +1,16 @@
 package javaTests.tests;
 
+import javaTests.steps.*;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Listeners;
 import org.testng.annotations.Test;
-
-import javaTests.steps.AddDeviceSteps;
-import javaTests.steps.DeviceListSteps;
-import javaTests.steps.FooterSteps;
-import javaTests.steps.HouseSteps;
-import javaTests.steps.SpinnerSteps;
 import utils.BaseAction;
 import utils.BaseTest;
 import utils.Listener.Listener;
 import utils.data.Strings;
 
 @Listeners(Listener.class)
-public class CreateNewHouseTest extends BaseTest {
+public class CreateNewHouseWifiTest extends BaseTest {
     private SpinnerSteps spinner;
     private HouseSteps house;
     private BaseAction baseAction;
@@ -34,7 +29,7 @@ public class CreateNewHouseTest extends BaseTest {
     }
 
     @Test
-    public void createNewHouseTest() throws InterruptedException {
+    public void createNewHouseWifi() throws InterruptedException {
 
         System.out.println("Create new house for WIFI");
         spinner.clickSpinner();
@@ -80,43 +75,6 @@ public class CreateNewHouseTest extends BaseTest {
         spinner.clickSpinner();
         spinner.checkHouseListWiFi();
         baseAction.pressBack();
-
-        System.out.println("Create new house for CC");
-        spinner.clickSpinner();
-        spinner.checkEnterByInviteText();
-        spinner.clickCreateNewHouseBtn();
-        house.checkCreateNewHouseTitle();
-        house.checkNewHouseNameDescriptionText();
-        house.checkCreateHouseBtnText();
-        house.checkEnterByInviteBtnText();
-        house.clickHelpBtn();
-        Thread.sleep(2000);
-        baseAction.pressBack();
-        house.clickCreateHouseBtn();
-        house.checkErrorEmptyNameNewHouseText();
-        house.checkErrorEmptyNameNewHouseBtnText();
-        house.clickOkBtnError();
-        house.clickEnterByInviteBtn();
-        baseAction.pressBack();
-        house.inputNewHouseName(Strings.HOUSE_NAME_CC);
-        baseAction.hideKeyBoard();
-        house.clickCreateHouseBtn();
-        house.checkCreateNewHouseTitle();
-        house.checkGotoAddDevicesBtnText();
-        house.checkContinueBtnText();
-        house.checkCongratulationMessageText();
-        house.checkDescriptionText();
-        house.clickHelpBtn();
-        Thread.sleep(2000);
-        baseAction.pressBack();
-        house.clickGoToAddDevicesBtn();
-        addDevice.checkAddDeviceGroupEmptyHouse();
-        baseAction.pressBack();
-        spinner.checkCurrentHouseCC();
-        deviceList.checkNotAvailableDevicesText();
-        spinner.clickSpinner();
-        spinner.checkHouseListWiFi();
-        spinner.checkHouseListCC();
 
     }
 }
