+++ src/test/java/utils/data/Strings.java
@@ -9,8 +9,8 @@ public class Strings {
     public final static String INCORRECT_FULL_PASSWORD = "111111";
     public final static String INCORRECT_SMSCODE = "1111";
     public final static String NOT_FULL_SMSCODE = "11";
-    public final static String HOUSE_NAME_WIFI = "Wi-fi Test_Name";
-    public final static String HOUSE_NAME_CC = "CC Test_Name";
+    public final static String HOUSE_NAME_WIFI = "Wi-fi test";
+    public final static String HOUSE_NAME_CC = "CC test";
     public final static String WIFI_PASSWORD = "qakRgCMr";
 
     //Default device_name
